+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - remove unused import from security system component
 - optimise api calls on opening uninstall component
+- dependency updates
 
 ### Homebridge Dependencies